@@ -1,71 +1,58 @@
 {
-  "name": "automacao-cancelamento",
-  "version": "1.0.0",
-  "main": "main.js",
-  "description": "Aplicação de Automação para Cancelamento",
-  "author": "Seu Nome",
-  "license": "ISC",
+  "name": "vite-react-typescript-starter",
+  "private": true,
+  "version": "0.0.0",
+  "type": "module",
   "scripts": {
-    "test": "echo \"Error: no test specified\" && exit 1",
-    "start": "chcp 65001 && electron .",
-    "electron": "electron .",
-    "build": "electron-builder"
+    "dev": "vite",
+    "build": "vite build",
+    "lint": "eslint .",
+    "preview": "vite preview",
+    "typecheck": "tsc --noEmit -p tsconfig.app.json",
+    "start:backend": "node seu_servidor_node.js",
+    "start": "node index.js",
+    "electron:dev": "cross-env VITE_DEV_SERVER_URL=http://localhost:5173 NODE_ENV=development electron .",
+    "dev:all": "concurrently \"npm:dev\" \"npm:electron:dev\""
   },
-  "keywords": [
-    "automação",
-    "cancelamento",
-    "electron"
-  ],
   "dependencies": {
     "@google-cloud/translate": "^9.0.1",
+    "@supabase/supabase-js": "^2.57.4",
     "axios": "^1.9.0",
     "chalk": "^4.1.2",
     "date-fns": "^4.1.0",
     "dotenv": "^16.4.7",
     "express": "^4.21.2",
+    "cors": "^2.8.5",
     "franc-min": "^6.2.0",
     "google-auth-library": "^9.15.1",
     "google-spreadsheet": "^4.1.4",
+    "googleapis": "^164.1.0",
+    "lucide-react": "^0.344.0",
+    "open": "^10.2.0",
     "openai": "^4.87.3",
+    "react": "^18.3.1",
+    "react-dom": "^18.3.1",
     "winston": "^3.17.0",
     "winston-daily-rotate-file": "^5.0.0",
     "xterm": "^5.3.0"
   },
   "devDependencies": {
-    "electron": "^35.1.5",
-    "electron-builder": "^26.0.12"
-  },
-  "build": {
-    "appId": "com.seu.nome.automacao",
-    "productName": "Automacao Cancelamento",
-    "files": [
-      "**/*"
-    ],
-    "extraFiles": [
-      {
-        "from": ".env",
-        "to": ".env"
-      }
-    ],
-    "directories": {
-      "buildResources": "src/assets"
-    },
-    "win": {
-      "icon": "capa.ico",
-      "target": "nsis"
-    },
-    "nsis": {
-      "oneClick": false,
-      "perMachine": false,
-      "allowToChangeInstallationDirectory": true
-    },
-    "mac": {
-      "icon": "capa.icns",
-      "target": "dmg"
-    },
-    "linux": {
-      "icon": "capa.png",
-      "target": "AppImage"
-    }
+    "@eslint/js": "^9.9.1",
+    "@types/react": "^18.3.5",
+    "@types/react-dom": "^18.3.0",
+    "@vitejs/plugin-react": "^4.3.1",
+    "autoprefixer": "^10.4.18",
+    "concurrently": "^9.2.1",
+    "cross-env": "^10.1.0",
+    "electron": "^38.4.0",
+    "eslint": "^9.9.1",
+    "eslint-plugin-react-hooks": "^5.1.0-rc.0",
+    "eslint-plugin-react-refresh": "^0.4.11",
+    "globals": "^15.9.0",
+    "postcss": "^8.4.35",
+    "tailwindcss": "^3.4.1",
+    "typescript": "^5.5.3",
+    "typescript-eslint": "^8.3.0",
+    "vite": "^5.4.2"
   }
 }
