@@ -1,26 +1,23 @@
-const winston = require('winston');
-const path = require('path');
-const { app } = require('electron');
-const fs = require('fs');
-const DailyRotateFile = require('winston-daily-rotate-file');
+import winston from 'winston';
+import path from 'path';
+import fs from 'fs';
+import DailyRotateFile from 'winston-daily-rotate-file';
 
-// Variável para armazenar a janela do Electron que receberá os logs
-let sendToWindow = null;
-
-// Diretório onde os logs serão armazenados
-const logDir = path.join(app.getPath('userData'), 'logs');
-
-// Garante que o diretório existe
+// Tenta importar 'app' do Electron, se disponível
+let logDir;
+try {
+  const { app } = await import('electron');
+  logDir = path.join(app.getPath('userData'), 'logs');
+} catch (e) {
+  // Fallback para pasta local se Electron não estiver disponível
+  logDir = path.join(process.cwd(), 'logs');
+}
 if (!fs.existsSync(logDir)) {
   fs.mkdirSync(logDir, { recursive: true });
 }
-
-// Formato do log
 const logFormat = winston.format.printf(({ level, message, timestamp }) => {
   return `${timestamp} [${level.toUpperCase()}]: ${message}`;
 });
-
-// Criando o logger com Winston
 const logger = winston.createLogger({
   level: 'info',
   format: winston.format.combine(
@@ -28,22 +25,14 @@ const logger = winston.createLogger({
     logFormat,
   ),
   transports: [
-    new DailyRotateFile({
-      filename: path.join(logDir, 'app-%DATE%.log'),
-      datePattern: 'YYYY-MM-DD',
-      maxSize: '15m',
-      maxFiles: '7d',
-    }),
+    // Removido: gravação automática do app-%DATE%.log
     new winston.transports.Console(),
   ],
 });
-
-// 👉 Função para definir a janela do Electron que receberá os logs
+let sendToWindow = null;
 function setSender(window) {
   sendToWindow = window;
 }
-
-// 👉 Função genérica para enviar logs para o frontend
 function sendToRenderer(level, message) {
   if (sendToWindow && sendToWindow.webContents) {
     sendToWindow.webContents.send(
@@ -51,30 +40,96 @@ function sendToRenderer(level, message) {
       `[${level.toUpperCase()}] ${message}`,
     );
   } else {
-    console.log(`[${level.toUpperCase()}] ${message}`); // Se a janela não estiver disponível, loga no console
+    // No-op in backend puro
   }
 }
-
-// 👉 Sobrescrevendo os métodos padrão do logger para também enviar ao frontend
 ['info', 'warn', 'error', 'debug'].forEach((level) => {
   const original = logger[level].bind(logger);
   logger[level] = (...args) => {
     const message = args
       .map((arg) => (typeof arg === 'string' ? arg : JSON.stringify(arg)))
       .join(' ');
-
     original(message);
     sendToRenderer(level, message);
+    // Adiciona ao log do ticket se estiver processando
+    if (typeof global !== 'undefined' && global.appendTicketLog) {
+      global.appendTicketLog(
+        `${new Date().toISOString()} [${level.toUpperCase()}]: ${message}`,
+      );
+    }
   };
 });
-
-// 👉 log() = atalho para logger.info()
+// Garante que appendTicketLog esteja disponível globalmente
+if (typeof global !== 'undefined') {
+  global.appendTicketLog = appendTicketLog;
+}
 function log(...args) {
   logger.info(...args);
 }
 
-module.exports = {
+// --- LOG DE TICKETS INDIVIDUAIS ---
+let currentTicketLog = null;
+let currentTicketId = null;
+let currentTicketUser = null;
+let currentTicketOrderIds = null;
+let currentTicketLogs = [];
+
+function startTicketLog(ticketId, user, orderIds) {
+  if (currentTicketLog) finishTicketLog();
+  currentTicketId = ticketId;
+  currentTicketUser = user;
+  currentTicketOrderIds = orderIds;
+  currentTicketLogs = [];
+}
+
+function appendTicketLog(message) {
+  if (currentTicketId) {
+    // Formata data/hora para o Brasil (DD/MM/YYYY HH:mm:ss)
+    const now = new Date();
+    const brDate = now.toLocaleString('pt-BR', {
+      timeZone: 'America/Sao_Paulo',
+      hour12: false,
+    });
+    currentTicketLogs.push({ timestamp: brDate, message });
+  }
+}
+
+function finishTicketLog() {
+  if (!currentTicketId) return;
+  const logText = currentTicketLogs
+    .map(
+      (l) =>
+        `[${l.timestamp}] ${l.message.replace(
+          /^\d{4}-\d{2}-\d{2}T.*?\[([A-Z]+)\]:/,
+          '[$1]:',
+        )}`,
+    )
+    .join('\n');
+  const now = new Date();
+  const dateFolder = `${String(now.getDate()).padStart(2, '0')}-${String(
+    now.getMonth() + 1,
+  ).padStart(2, '0')}-${now.getFullYear()}`;
+  const logDir = path.join(process.cwd(), 'logs', 'tickets', dateFolder);
+  if (!fs.existsSync(logDir)) fs.mkdirSync(logDir, { recursive: true });
+  const filePathLog = path.join(logDir, `ticket_${currentTicketId}.log`);
+  fs.writeFileSync(filePathLog, logText, 'utf-8');
+  // Salvamento local apenas (Google Drive removido)
+  try {
+    logger.info(`✅ Log salvo localmente em ${filePathLog}`);
+  } catch (e) {
+    console.error('Erro ao registrar salvamento de log local:', e.message);
+  }
+  currentTicketId = null;
+  currentTicketUser = null;
+  currentTicketOrderIds = null;
+  currentTicketLogs = [];
+}
+
+export {
   setSender,
   log,
   logger,
+  startTicketLog,
+  appendTicketLog,
+  finishTicketLog,
 };
