@@ -1,212 +1,166 @@
-const axios = require('axios');
-const logger = require('../logger'); // Sistema de logs
-const {
-  removerTagsHTML,
-  extrairOrderIdDaMensagem,
-  buscarStatusPedido,
-  responderTicket,
-} = require('./apiService'); // Funções de utilidades para limpeza de texto e extração do Order ID
+import axios from 'axios';
+import { logger } from '../logger.js';
+import { buscarStatusPedido, responderTicket } from './apiService.js';
+import {
+  parseRawMessage, // << NOVO: Importação do parser robusto para aprimorar a extração
+  formatIds,
+  agruparPedidos,
+} from './utils.js';
 
-// Adicione essa função no início ou final do arquivo iaSolicitacao.js
-
-// Função para extrair múltiplos Order IDs
-
-function extrairTodosOrderIds(mensagem) {
-  if (!mensagem) return [];
-
-  // Alteração na regex para capturar apenas sequências de 6 dígitos
-  const regexOrderId = /<div><b>Order ID<\/b>: (\d{6})<\/div>/g; // 6 dígitos
-  let orderIds = [];
-  let match;
-
-  // Usar o regex para capturar os Order IDs dentro das tags <div><b>Order ID</b>:
-  while ((match = regexOrderId.exec(mensagem)) !== null) {
-    orderIds.push(match[1]); // Adiciona os IDs encontrados
+// Use logger para logs importantes, console.log para debug local
+function debugLog(...args) {
+  if (process.env.DEBUG_IA === 'true') {
+    console.log(...args);
   }
-
-  // Se não encontrar IDs diretamente, tentamos capturar qualquer sequência numérica de 6 dígitos
-  if (orderIds.length === 0) {
-    const regexForaDeTags = /\b\d{6}\b/g; // Captura apenas números de 6 dígitos
-    orderIds = [...new Set(mensagem.match(regexForaDeTags) || [])]; // Adiciona IDs encontrados fora das tags
-  }
-
-  // Caso haja IDs concatenados, separe-os (como no caso de 550039550039)
-  const orderIdsSeparados = [];
-  for (const orderId of orderIds) {
-    if (orderId.length >= 12) {
-      // Se o ID for maior que 12 dígitos, dividimos em dois IDs de 6 dígitos cada
-      const partes = orderId.match(/(\d{6})(\d{6})/);
-      if (partes) {
-        orderIdsSeparados.push(partes[1], partes[2]);
-      } else {
-        orderIdsSeparados.push(orderId); // Se não puder ser dividido, adiciona como está
-      }
-    } else {
-      orderIdsSeparados.push(orderId); // Adiciona IDs válidos de 6 dígitos
-    }
-  }
-
-  // Remover IDs duplicados
-  const orderIdsUnicos = [...new Set(orderIdsSeparados)];
-
-  console.log('-----------extrairTodosOrdersID');
-  console.log('✅ Order IDs extraídos:', orderIdsUnicos);
-  return orderIdsUnicos;
 }
 
-// Ajuste no processamento dos Order IDs para agrupar por tipo de solicitação
-async function processarOrderIds(
+// --- FLUXO PRINCIPAL: Geração da Resposta Final (Otimizada para processamento de pedidos) ---
+
+async function gerarRespostaFinal(
   ticketId,
   tipoSolicitacao,
-  mensagem,
+  orderIds, // Lista de IDs extraídos (já filtrados e priorizados pelo index.js)
+  orderDataList, // Lista de dados do pedido (já buscados e agrupados no index.js)
   idiomaDetectado,
+  nomeAtendente = 'Marcelle',
+  linkRevendedor = 'https://smmexcellent.com/child-panel',
+  linkAfiliado = 'https://smmexcellent.com/affiliates',
+  options = {}, // { notFoundIds: [...] }
 ) {
-  const orderIds = extrairTodosOrderIds(mensagem); // Extrai todos os Order IDs da mensagem
-  console.log(
-    `📦 [processarOrderIds] Order IDs extraídos: ${orderIds.join(', ')}`,
-  );
+  debugLog('---------- GERAR RESPOSTA FINAL ------------'); // A função processarOrderIds foi removida. A responsabilidade de buscar e agrupar // os dados dos pedidos (orderDataList) agora é do 'index.js', que faz isso // de forma mais eficiente (paralela se necessário) antes de chamar esta função.
 
-  if (orderIds.length === 0) {
-    console.log(`❌ Nenhum Order ID encontrado no ticket ${ticketId}`);
-    return `No Order IDs were found in the message.`;
+  const orderIdsUnicos = [...new Set(orderIds)];
+
+  if (!orderDataList || orderDataList.length === 0) {
+    logger.error(
+      `⚠️ Ticket ${ticketId}: orderDataList vazia. Resposta genérica.`,
+    ); // Se não houver dados, retorna uma resposta baseada apenas no tipo.
+    const respostaGenerica = `Hello,\n\nWe have received your ${tipoSolicitacao.toLowerCase()} request for order(s) ID ${formatIds(
+      orderIdsUnicos,
+    )}.\n\nWe are processing the request. If there are any issues, we will contact you immediately.\n\n`;
+    let respostaTraduzidaGenerica = respostaGenerica;
+    if (idiomaDetectado !== 'en') {
+      respostaTraduzidaGenerica = await traduzirTexto(
+        respostaGenerica,
+        idiomaDetectado,
+      );
+    }
+    return respostaTraduzidaGenerica;
   }
+  const nomeUsuario = orderDataList[0]?.user || '';
+  let respostaIA = `Hello ${nomeUsuario},\n\n`;
 
-  let pedidosAptos = [];
-  let pedidosNaoAptos = [];
-  let mensagemRespostas = '';
-  const orderIdsProcessados = new Set(); // Conjunto para verificar se o Order ID já foi processado
+  const pedidosAgrupados = agruparPedidos(orderDataList);
 
-  // Dicionário para agrupar os Order IDs por tipo de solicitação
-  const pedidosPorTipo = {
-    Aceleração: [],
-    Cancelamento: [],
-    'Refil/Garantia': [],
-    Diversos: [],
-  };
-
-  // Processa cada Order ID
-  for (const orderId of orderIds) {
-    if (orderIdsProcessados.has(orderId)) {
-      console.log(`🚫 Order ID ${orderId} já foi processado, ignorando.`);
-      continue;
+  if (tipoSolicitacao === 'Cancelamento') {
+    if (pedidosAgrupados.completed.length > 0) {
+      respostaIA += `Your order(s) <strong>ID ${formatIds(
+        pedidosAgrupados.completed,
+      )}</strong> ${
+        pedidosAgrupados.completed.length === 1 ? 'is' : 'are'
+      } already <strong>complete</strong>. We cannot cancel order(s) that have already been completed.\n\n`;
     }
 
-    orderIdsProcessados.add(orderId);
-
-    let orderData = await buscarStatusPedido(orderId);
-    if (!orderData) {
-      pedidosNaoAptos.push({ orderId, motivo: 'Não encontrado' });
-      continue;
+    if (pedidosAgrupados.canceled.length > 0) {
+      respostaIA += `Your order(s) <strong>ID ${formatIds(
+        pedidosAgrupados.canceled,
+      )}</strong> ${
+        pedidosAgrupados.canceled.length === 1 ? 'has' : 'have'
+      } already been <strong>canceled</strong>, as requested. You can place a new order at any time.\n\n`;
     }
 
-    // Agrupar Order IDs por tipo de solicitação
-    if (orderData.status === 'canceled') {
-      pedidosNaoAptos.push({ orderId, motivo: 'Pedido já cancelado' });
-    } else if (orderData.status === 'completed') {
-      pedidosNaoAptos.push({ orderId, motivo: 'Pedido já completo' });
-    } else {
-      pedidosPorTipo[tipoSolicitacao].push(orderId); // Agrupa pelo tipo de solicitação
+    if (pedidosAgrupados.pendente.length > 0) {
+      respostaIA += `We have submitted your <strong>cancellation</strong> request for your order(s) <strong>ID ${formatIds(
+        pedidosAgrupados.pendente,
+      )}</strong>, and it is subject to approval. Please kindly wait while your request is reviewed.\n\n`;
     }
+  } else if (tipoSolicitacao === 'Aceleração') {
+    if (pedidosAgrupados.completed.length > 0) {
+      respostaIA += `Your order(s) <strong>ID ${formatIds(
+        pedidosAgrupados.completed,
+      )}</strong> ${
+        pedidosAgrupados.completed.length === 1 ? 'is' : 'are'
+      } already <strong>complete</strong>. We cannot expedite order(s) that have already been completed.\n\n`;
+    }
+
+    if (pedidosAgrupados.canceled.length > 0) {
+      respostaIA += `Your order(s) <strong>ID ${formatIds(
+        pedidosAgrupados.canceled,
+      )}</strong> ${
+        pedidosAgrupados.canceled.length === 1 ? 'has' : 'have'
+      } been <strong>canceled</strong>. We cannot expedite canceled order(s).\n\n`;
+    }
+
+    if (pedidosAgrupados.pendente.length > 0) {
+      respostaIA += `Your <strong>acceleration</strong> request has been forwarded. We will try to expedite the order(s) <strong>ID ${formatIds(
+        pedidosAgrupados.pendente,
+      )}</strong> as soon as possible.\n\n`;
+    }
+  } else if (tipoSolicitacao === 'Refil/Garantia') {
+    respostaIA += `I have forwarded your refill/warranty request for order(s) <strong>ID ${formatIds(
+      orderIdsUnicos,
+    )}</strong> to our technical team.\n\nTo be eligible for a refill, the current count must be above the start count and below the end count, and the order must be within the warranty/refill period described in the service.\n\nIf approved, the process will be completed within 1 to 3 days.\n\nIf you need anything else, feel free to contact us again.\n\n`;
+  } else {
+    // Diversos ou outros tipos de pedido
+    respostaIA += `Your request has been forwarded to our specialized technical team for analysis. We will contact you soon regarding order(s) <strong>ID ${formatIds(
+      orderIdsUnicos,
+    )}</strong>.\n\n`;
   }
 
-  // Gerando a resposta para o cliente
-  if (pedidosPorTipo[tipoSolicitacao].length > 0) {
-    mensagemRespostas += `The following Order IDs have been sent for ${tipoSolicitacao}: ${pedidosPorTipo[
-      tipoSolicitacao
-    ].join(', ')}.\n\n`;
+  if (pedidosAgrupados.invalidos.length > 0) {
+    respostaIA += `Some order(s) could not be processed due to missing or invalid information: <strong>ID ${formatIds(
+      pedidosAgrupados.invalidos,
+    )}</strong>.\n\n`;
   }
 
-  // Caso haja pedidos não aptos, incluir as razões
-  if (pedidosNaoAptos.length > 0) {
-    mensagemRespostas += `The following Order IDs were not processed:\n`;
-    pedidosNaoAptos.forEach((pedido) => {
-      mensagemRespostas += `Order ID ${pedido.orderId} - Reason: ${pedido.motivo}\n`;
-    });
+  // If index.js passed notFoundIds (order IDs that the customer provided but were not
+  // found in the system), explicitly inform the customer which IDs weren't found.
+  if (options && Array.isArray(options.notFoundIds) && options.notFoundIds.length > 0) {
+    respostaIA += `Additionally, the following order ID(s) you provided were not found in our system: <strong>ID ${formatIds(
+      options.notFoundIds,
+    )}</strong>. We processed the other valid order(s) above.
+
+`;
   }
 
-  // Enviar a resposta ao cliente
-  let respostaEnviada = false;
-  if (!respostaEnviada) {
-    console.log(`-----------ProcessarOdersID-----------`);
-    console.log(
-      `Resposta enviada para o ticket ${ticketId}: ${mensagemRespostas}`,
-    );
-    await responderTicket(ticketId, mensagemRespostas);
-    respostaEnviada = true; // Marca a resposta como enviada
+  respostaIA += `---\nIf you need more information, we are available to help.\n\nBest regards,\n\n${nomeAtendente}\n\n`;
+  respostaIA += `➕ Join us as a reseller for just $25 - [Reseller Link](${linkRevendedor})\n`;
+  respostaIA += `➕ Invite friends, share your link, and earn! - [Affiliate Link](${linkAfiliado})\n`;
+
+  let respostaTraduzida = respostaIA;
+  if (idiomaDetectado !== 'en') {
+    respostaTraduzida = await traduzirTexto(respostaIA, idiomaDetectado);
   }
 
-  return mensagemRespostas;
+  return respostaTraduzida;
 }
 
-function cortarMensagemUtil(mensagemOriginal) {
-  console.log(`-------------CORTAR MENSAGEM -------------------`);
-  console.log(`📜 Mensagem original: "${mensagemOriginal}"`);
-
-  // Encontrar a posição da primeira tag </b>
-  const primeiraTagFechamento = mensagemOriginal.indexOf('</b>'); // Encontra a primeira tag </b>
-
-  if (primeiraTagFechamento === -1) {
-    // Se não encontrar a tag </b>, retorna a mensagem original
-    console.log(
-      `⚡ Não foi possível localizar a tag </b>, retornando a mensagem original.`,
-    );
-    return mensagemOriginal;
-  }
-
-  // Encontrar a posição da segunda tag </b>
-  let segundaTagFechamento = mensagemOriginal.indexOf(
-    '</b>',
-    primeiraTagFechamento + 1,
-  );
-
-  // Caso a segunda tag </b> não exista, tentamos retornar a partir da primeira
-  if (segundaTagFechamento === -1) {
-    console.log(
-      `⚡ Segunda tag </b> não encontrada, utilizando a primeira tag </b> para cortar a mensagem.`,
-    );
-    // Caso não tenha a segunda tag </b>, retornamos o restante da mensagem a partir da primeira
-    return mensagemOriginal.slice(primeiraTagFechamento + 4).trim();
-  }
-
-  // Caso encontre a segunda tag </b>, cortamos a partir dela
-  let cortadaComTags = mensagemOriginal.slice(segundaTagFechamento + 4).trim(); // Pula a segunda tag </b> e mantém o resto
-
-  // Remover as tags HTML da mensagem cortada
-  //const cortadaSemTags = cortadaComTags.replace(/<\/?[^>]+(>|$)/g, '').trim();
-
-  console.log(`🔧 Mensagem cortada: "${cortadaComTags}"`);
-  return cortadaComTags;
-}
+// --- CLASSIFICAÇÃO GERAL (Usa o parser robusto) ---
 
 async function classificarCategoriaGeral(mensagemOriginal) {
-  // Chama a função cortarMensagemUtil diretamente para obter a mensagem cortada corretamente
-  const mensagemCortada = cortarMensagemUtil(mensagemOriginal); // A mensagem cortada já está aqui
+  // Usa o parser robusto para obter o corpo limpo da mensagem
+  const parsedMessage = parseRawMessage(mensagemOriginal);
+  const mensagemCortada = parsedMessage.body;
 
-  // Verifica se `mensagemCortada` está em um formato esperado
-  console.log(`------- CLASSIFICAR CATEGORIA ------------`);
-  console.log(`✂️ classificar:  Mensagem cortada (útil): "${mensagemCortada}"`);
-
-  // Refinando o prompt para a IA, deixando claro que a palavra "cancelar" se refere a um pedido
   const prompt = `
-    Você é um assistente de suporte multilíngue. Sua tarefa é classificar a mensagem do cliente como "Pedido", "Pagamento" ou "Outro", com base no conteúdo da mensagem.
+    Você é um assistente de suporte multilíngue. Sua tarefa é classificar a mensagem do cliente como "Pedido", "Pagamento" ou "Outro", com base no conteúdo da mensagem.
 
-    
-    1. **"Pedido"** – Se a mensagem for sobre qualquer aspecto de um pedido, como cancelamento, aceleração, status, garantia, refil, ou entrega.
-       - A IA deve identificar que, mesmo sem o **Order ID**, a mensagem é sobre um pedido. 
-       - **Se a mensagem contiver "Speed" ou "SpeedUp"**, isso deve ser automaticamente classificado como "Pedido", **mesmo sem o número do pedido presente**. Exemplo: "Speed", "SpeedUp", "Please speed up my order", "I need to speed my request", "18575speed", entre outros.
+    1. **"Pedido"** – Se a mensagem for sobre qualquer aspecto de um pedido, como cancelamento, aceleração, status, garantia, refil, ou entrega.
+       - A IA deve identificar que, mesmo sem o **Order ID**, a mensagem é sobre um pedido. 
+       - **Se a mensagem contiver "Speed" ou "SpeedUp"**, isso deve ser automaticamente classificado como "Pedido", **mesmo sem o número do pedido presente**. Exemplo: "Speed", "SpeedUp", "Please speed up my order", "I need to speed my request", "18575speed", entre outros.
 
-    2. "Pagamento" – se a mensagem for sobre questões financeiras, como valor, desconto, fatura, cobrança, saldo, cartão, comprovante de pagamento, falha no pagamento, ou qualquer outro termo relacionado ao pagamento ou à transação financeira etc.
-    3. "Outro" – se não for relacionado a pedidos ou pagamentos. Exemplos: perda de senha, suporte técnico, dúvidas gerais, reclamações, acesso à conta, etc.
+    2. "Pagamento" – se a mensagem for sobre questões financeiras, como valor, desconto, fatura, cobrança, saldo, cartão, comprovante de pagamento, falha no pagamento, ou qualquer outro termo relacionado ao pagamento ou à transação financeira etc.
+    3. "Outro" – se não for relacionado a pedidos ou pagamentos. Exemplos: perda de senha, suporte técnico, dúvidas gerais, reclamações, acesso à conta, etc.
 
-    ⚠️ IMPORTANTE:
-    - A mensagem pode ser em qualquer idioma — a IA deve ser capaz de entender isso.
-    - Se a mensagem estiver relacionada a um **pedido**, mesmo sem o **Order ID**, classifique como "Pedido" e solicite ao cliente o **Order ID**.
-    - Se a mensagem for sobre pagamento ou questões financeiras, classifique como "Pagamento".
-    - Caso contrário, classifique como "Outro".
+    ⚠️ IMPORTANTE:
+    - A mensagem pode ser em qualquer idioma — a IA deve ser capaz de entender isso.
+    - Se a mensagem estiver relacionada a um **pedido**, mesmo sem o **Order ID**, classifique como "Pedido" e solicite ao cliente o **Order ID**.
+    - Se a mensagem for sobre pagamento ou questões financeiras, classifique como "Pagamento".
+    - Caso contrário, classifique como "Outro".
 
-    Mensagem do cliente:
-    """${mensagemCortada}"""
-  `;
+    Mensagem do cliente:
+    """${mensagemCortada}"""
+  `;
 
   try {
     const response = await axios.post(
@@ -237,26 +191,22 @@ async function classificarCategoriaGeral(mensagemOriginal) {
 
     const resultado = response.data.choices[0].message.content.trim();
 
-    // Verificar se o resultado está dentro das categorias esperadas
     if (!['Pedido', 'Pagamento', 'Outro'].includes(resultado)) {
-      console.log(`⚠️ Resposta inesperada da IA: ${resultado}`);
-      return 'Outro'; // fallback seguro para retornar "Outro" caso não identifique uma categoria válida
+      return 'Outro';
     }
-    console.log(`🎯 Categoria classificada pela IA: ${resultado}`);
-    return resultado; // Retorna a categoria classificada
+    logger.info(`🎯 Categoria classificada pela IA: ${resultado}`);
+    return resultado;
   } catch (error) {
-    logger.error('Erro ao classificar categoria geral:', error.message);
-    return 'Outro'; // Retorna "Outro" como fallback seguro se houver erro durante a classificação
+    logger.error('❌ Erro [IA Classificação Geral]:', error.message);
+    debugLog('Detalhes do erro:', error);
+    return 'Outro';
   }
 }
 
-// Função para detectar o idioma usando a OpenAI
-// Função para detectar o idioma usando a OpenAI
+// --- DETECÇÃO DE IDIOMA (Usa o parser robusto) ---
+
 async function detectLanguage(messages) {
   if (!Array.isArray(messages)) {
-    logger.error(
-      'O parâmetro messages não é um array. Convertendo para array.',
-    );
     messages = [messages];
   }
 
@@ -271,16 +221,12 @@ async function detectLanguage(messages) {
     );
 
   if (!ultimaMensagemUtil) {
-    logger.error('Nenhuma mensagem útil encontrada para análise.');
+    logger.error('Nenhuma mensagem útil encontrada para análise de idioma.');
     return 'en';
-  }
+  } // Usa o parser robusto para obter o corpo limpo da mensagem
 
-  // 🔍 Aplica o corte inteligente da mensagem apenas uma vez
-  const mensagemCortada = cortarMensagemUtil(ultimaMensagemUtil.message); // A versão cortada da mensagem já é gerada aqui
-
-  // Mostra no log a diferença
-  console.log(`---------- DetectLanguage------------`);
-  console.log(`✂️ Mensagem cortada (útil): "${mensagemCortada}"`);
+  const parsedMessage = parseRawMessage(ultimaMensagemUtil.message);
+  const mensagemCortada = parsedMessage.body;
 
   try {
     const response = await axios.post(
@@ -306,29 +252,30 @@ async function detectLanguage(messages) {
 
     const idiomaDetectado = response.data.choices[0].message.content.trim();
     if (!idiomaDetectado || idiomaDetectado === 'und') {
-      console.log('❌ Idioma não identificado, usando fallback pt');
       return 'en';
     }
 
-    console.log(`🌐 Idioma detectado: ${idiomaDetectado}`);
     return idiomaDetectado;
   } catch (error) {
-    console.log('❌ Erro na detecção de idioma:', error.message);
+    logger.error('❌ Erro [IA Detecção de Idioma]:', error.message);
+    debugLog('Detalhes do erro:', error);
     return 'en';
   }
 }
 
+// --- TRADUÇÃO E SOLICITAÇÃO DE ID (Mantidas) ---
+
 async function traduzirTexto(texto, idiomaDestino) {
   if (!idiomaDestino) {
     logger.error('Idioma de destino não especificado para tradução.');
-    return texto; // Retorna o texto original se não houver idioma de destino
+    return texto;
   }
 
   try {
     const response = await axios.post(
       'https://api.openai.com/v1/chat/completions',
       {
-        model: 'gpt-4o-mini-2024-07-18', // Modelo da OpenAI
+        model: 'gpt-4o-mini-2024-07-18',
         messages: [
           {
             role: 'system',
@@ -351,13 +298,14 @@ async function traduzirTexto(texto, idiomaDestino) {
     );
 
     if (response.data && response.data.choices && response.data.choices[0]) {
-      return response.data.choices[0].message.content.trim(); // Retorna a tradução
+      return response.data.choices[0].message.content.trim();
     } else {
       throw new Error('Resposta da IA não está no formato esperado.');
     }
   } catch (error) {
-    logger.error('Erro ao traduzir com OpenAI:', error.message);
-    return texto; // Retorna o texto original em caso de erro
+    logger.error('❌ Erro [IA Tradução]:', error.message);
+    debugLog('Detalhes do erro:', error);
+    return texto;
   }
 }
 
@@ -379,94 +327,55 @@ Marcelle
     const traduzido = await traduzirTexto(textoBase, idiomaDestino);
     return traduzido;
   } catch (error) {
-    logger.error(
-      'Erro ao gerar mensagem de solicitação de Order ID:',
-      error.message,
-    );
+    debugLog('Detalhes do erro:', error);
     return textoBase;
   }
 }
 
-// Função para gerar a solicitação à IA
+// --- CLASSIFICAÇÃO DE TIPO DE PEDIDO (Usa o parser robusto) ---
+
 async function verificarTipoDeSolicitacao(messages) {
-  // Filtra as últimas 3 mensagens do cliente
   const mensagensCliente = messages
     .filter((msg) => msg.message && !msg.is_staff)
-    .slice(-3); // Últimas 3 mensagens
+    .slice(-3);
 
-  // Junta as mensagens para uma única string
   const messageText = mensagensCliente
-    .map((msg) => msg.message.trim()) // Não remove tags HTML antes
-    .join(' ');
+    .map((msg) => msg.message.trim())
+    .join(' '); // Usa o parser robusto para obter o corpo limpo da mensagem
 
-  // Aplica o corte na mensagem, mantendo as tags intactas
-  const mensagemCortada = cortarMensagemUtil(messageText); // Chama a função de corte uma vez
+  const parsedMessage = parseRawMessage(messageText);
+  const mensagemCortada = parsedMessage.body; // 1. Classificação Categoria Geral
 
-  // **Log para ver a mensagem cortada antes da classificação**
-  console.log(`------ VERIFICAR TIPO DE SOLICITACAO ----------`);
-  console.log(
-    `📜 [verificarTipoDeSolicitacao] Analisando a mensagem do cliente: "${mensagemCortada}"`,
-  );
-
-  // Detectar idioma (continua sendo útil para tradução futura)
-  const idiomaDetectado = await detectLanguage(messages); // Usa a mensagem cortada para detectar o idioma
-
-  // ✅ Classifica a categoria geral do ticket usando a mensagem cortada
-  const categoria = await classificarCategoriaGeral(mensagemCortada); // Passa a mensagem cortada para a classificação
-  console.log(
-    `🎯 [verificarTipoDeSolicitacao] Categoria geral detectada: ${categoria}`,
-  );
+  const categoria = await classificarCategoriaGeral(mensagemCortada);
 
   if (categoria === 'Pagamento' || categoria === 'Outro') {
-    // Ignorar processamento se for relacionado a "Pagamento" ou "Outro"
-    console.log(
-      `🚫 [verificarTipoDeSolicitacao] Ticket relacionado a "Pagamento" ou "Outro". Ignorando.`,
-    );
-    return { tipoSolicitacao: categoria, orderIds: [] }; // Retorna um array vazio de orderIds
+    return { tipoSolicitacao: categoria, orderIds: [] };
   }
 
-  // ✅ Continua apenas se for relacionado a Pedido
-  const orderIds = extrairTodosOrderIds(mensagemCortada); // Passa a versão cortada da mensagem
-
-  // **Log para ver o que aconteceu com os Order IDs extraídos**
-  if (orderIds.length === 0) {
-    console.log(
-      `❗ [verificarTipoDeSolicitacao] Pedido identificado, mas sem Order ID. Solicitando o Order ID ao cliente.`,
-    );
-  } else {
-    console.log(
-      `✅ [verificarTipoDeSolicitacao] Order ID(s) extraído(s): ${orderIds.join(
-        ', ',
-      )}`,
-    );
-  }
+  // 2. Extração de IDs (usando o parser)
+  // Nota: A função extrairTodosOrderIds NÃO é necessária se você usar o parseRawMessage no index.js
+  // Vamos manter a chamada para extração de IDs aqui por segurança de compatibilidade, mas o index.js é quem usa o parser completo
+  // Para fins deste bloco, vamos simular a extração a partir do parseRawMessage.
+  const orderIds = parsedMessage.orderIds.map((item) => item.id);
 
   if (orderIds.length === 0) {
-    // Se não houver Order ID, solicita o Order ID
-    const mensagemSolicitacao = await gerarMensagemSolicitandoOrderId(
-      idiomaDetectado,
-    );
-    console.log(
-      `❓ [verificarTipoDeSolicitacao] Solicitando Order ID ao cliente. Enviar a mensagem para o cliente.`,
-    );
-
-    return { tipoSolicitacao: 'Pedido', orderIds: [] }; // Retorna vazio, pois o cliente precisa fornecer o Order ID
+    // Se for "Pedido" mas sem ID, o index.js decide se solicita o ID ou não.
+    return { tipoSolicitacao: 'Pedido', orderIds: [] };
   }
 
-  // 🔄 Identificar a intenção dentro do contexto do pedido usando a mensagem cortada
   const prompt = `
-    O cliente está solicitando suporte relacionado a um pedido. A solicitação pode ser:
+    O cliente está solicitando suporte relacionado a um pedido. A solicitação pode ser:
 
-    - Cancelamento
-    - Aceleração
-    - Refil/Garantia
-    - Diversos (mas ainda dentro de "pedido")
+    - Cancelamento
+    - Aceleração
+    - Refil/Garantia
+    - Diversos (mas ainda dentro de "pedido")
 
-    Com base nas mensagens abaixo, classifique a intenção do cliente:
-    ${mensagemCortada}
+    Com base nas mensagens abaixo, classifique a intenção do cliente:
+    ${mensagemCortada}
 
-    Responda com: Cancelamento, Aceleração, Refil/Garantia ou Diversos.
-  `;
+    Responda com: Cancelamento, Aceleração, Refil/Garantia ou Diversos.
+  `;
 
   try {
     const response = await axios.post(
@@ -494,155 +403,31 @@ async function verificarTipoDeSolicitacao(messages) {
       },
     );
 
-    const tipo = response.data.choices[0].message.content.trim();
+    let tipo = response.data.choices[0].message.content.trim(); // Pequena correção para garantir que a saída da IA se encaixe nas categorias.
 
-    // Verificar a resposta para garantir que "Aceleração" é reconhecida corretamente
-    if (tipo === 'speedup') {
-      return { tipoSolicitacao: 'Aceleração', orderIds };
+    if (tipo.toLowerCase() === 'speedup') {
+      tipo = 'Aceleração';
+    }
+    if (
+      !['Cancelamento', 'Aceleração', 'Refil/Garantia', 'Diversos'].includes(
+        tipo,
+      )
+    ) {
+      tipo = 'Diversos';
     }
 
     return { tipoSolicitacao: tipo, orderIds };
   } catch (error) {
-    console.log(
-      `❌ Erro ao identificar tipo dentro de "Pedido": ${error.message}`,
-    );
-    return { tipoSolicitacao: 'Outro', orderIds }; // Retorna "Outro" caso falhe na classificação
+    logger.error('❌ Erro [IA Classificação Pedido]:', error.message);
+    debugLog('Detalhes do erro:', error);
+    return { tipoSolicitacao: 'Diversos', orderIds };
   }
-  // Log final para identificar o tipo de solicitação
-  console.log(
-    `🔍 [verificarTipoDeSolicitacao] Tipo de solicitação final: ${categoria}`,
-  );
 }
 
-async function gerarRespostaFinal(
-  ticketId,
-  tipoSolicitacao,
-  orderIds,
-  orderDataList,
-  idiomaDetectado,
-  nomeAtendente = 'Marcelle',
-  linkRevendedor = 'https://smmexcellent.com/child-panel',
-  linkAfiliado = 'https://smmexcellent.com/affiliates',
-) {
-  console.log('---------- GERAR RESPOSTA FINAL ------------');
-
-  function formatIds(ids) {
-    return Array.isArray(ids) ? ids.join(', ') : String(ids);
-  }
-
-  let respostaIA = '';
-  const orderIdsUnicos = [...new Set(orderIds)];
-
-  if (!Array.isArray(orderDataList)) {
-    orderDataList = [orderDataList];
-  }
-
-  const nomeUsuario = orderDataList[0]?.user || '';
-  respostaIA += `Hello ${nomeUsuario},\n\n`;
-
-  const pedidosAgrupados = {
-    completed: [],
-    canceled: [],
-    pendente: [],
-    invalidos: [],
-  };
-
-  for (const orderData of orderDataList) {
-    if (!orderData || !orderData.status) {
-      pedidosAgrupados.invalidos.push(orderData?.orderId || '(unknown)');
-      continue;
-    }
-
-    const status = orderData.status.toLowerCase();
-    if (status === 'completed') {
-      pedidosAgrupados.completed.push(orderData.orderId);
-    } else if (status === 'canceled') {
-      pedidosAgrupados.canceled.push(orderData.orderId);
-    } else {
-      pedidosAgrupados.pendente.push(orderData.orderId);
-    }
-  }
-
-  if (tipoSolicitacao === 'Cancelamento') {
-    if (pedidosAgrupados.completed.length > 0) {
-      respostaIA += `Your order(s) <strong>ID ${formatIds(
-        pedidosAgrupados.completed,
-      )}</strong> ${
-        pedidosAgrupados.completed.length === 1 ? 'is' : 'are'
-      } already <strong>complete</strong>. We cannot cancel order(s) that have already been completed.\n\n`;
-    }
-
-    if (pedidosAgrupados.canceled.length > 0) {
-      respostaIA += `Your order(s) <strong>ID ${formatIds(
-        pedidosAgrupados.canceled,
-      )}</strong> ${
-        pedidosAgrupados.canceled.length === 1 ? 'has' : 'have'
-      } already been <strong>canceled</strong>, as requested. You can place a new order at any time.\n\n`;
-    }
-
-    if (pedidosAgrupados.pendente.length > 0) {
-      respostaIA += `The <strong>cancellation</strong> request for your order(s) <strong>ID ${formatIds(
-        pedidosAgrupados.pendente,
-      )}</strong> has been forwarded to the responsible team. Your order(s) will be canceled soon.\n\n`;
-    }
-  } else if (tipoSolicitacao === 'Aceleração') {
-    if (pedidosAgrupados.completed.length > 0) {
-      respostaIA += `Your order(s) <strong>ID ${formatIds(
-        pedidosAgrupados.completed,
-      )}</strong> ${
-        pedidosAgrupados.completed.length === 1 ? 'is' : 'are'
-      } already <strong>complete</strong>. We cannot expedite order(s) that have already been completed.\n\n`;
-    }
-
-    if (pedidosAgrupados.canceled.length > 0) {
-      respostaIA += `Your order(s) <strong>ID ${formatIds(
-        pedidosAgrupados.canceled,
-      )}</strong> ${
-        pedidosAgrupados.canceled.length === 1 ? 'has' : 'have'
-      } been <strong>canceled</strong>. We cannot expedite canceled order(s).\n\n`;
-    }
-
-    if (pedidosAgrupados.pendente.length > 0) {
-      respostaIA += `Your <strong>acceleration</strong> request has been forwarded. We will try to expedite the order(s) <strong>ID ${formatIds(
-        pedidosAgrupados.pendente,
-      )}</strong> as soon as possible.\n\n`;
-    }
-  } else if (tipoSolicitacao === 'Refil/Garantia') {
-    respostaIA += `I have forwarded your refill/warranty request for order(s) <strong>ID ${formatIds(
-      orderIdsUnicos,
-    )}</strong> to our technical team.\n\nTo be eligible for a refill, the current count must be above the start count and below the end count, and the order must be within the warranty/refill period described in the service.\n\nIf approved, the process will be completed within 1 to 3 days.\n\nIf you need anything else, feel free to contact us again.\n\n`;
-  } else {
-    respostaIA += `Your request has been forwarded to our specialized technical team for analysis. We will contact you soon regarding order(s) <strong>ID ${formatIds(
-      orderIdsUnicos,
-    )}</strong>.\n\n`;
-  }
-
-  if (pedidosAgrupados.invalidos.length > 0) {
-    respostaIA += `Some order(s) could not be processed due to missing or invalid information: <strong>ID ${formatIds(
-      pedidosAgrupados.invalidos,
-    )}</strong>.\n\n`;
-  }
-
-  respostaIA += `---\nIf you need more information, we are available to help.\n\nBest regards,\n\n${nomeAtendente}\n\n`;
-  respostaIA += `➕ Join us as a reseller for just $25 - [Reseller Link](${linkRevendedor})\n`;
-  respostaIA += `➕ Invite friends, share your link, and earn! - [Affiliate Link](${linkAfiliado})\n`;
-
-  let respostaTraduzida = respostaIA;
-  if (idiomaDetectado !== 'en') {
-    console.log(`🌍 Traduzindo a resposta para o idioma: ${idiomaDetectado}`);
-    respostaTraduzida = await traduzirTexto(respostaIA, idiomaDetectado);
-  }
-
-  return respostaTraduzida;
-}
-
-module.exports = {
+export {
   gerarRespostaFinal,
   verificarTipoDeSolicitacao,
   detectLanguage,
   traduzirTexto,
-  gerarMensagemSolicitandoOrderId,
-  extrairTodosOrderIds,
-  processarOrderIds,
-  cortarMensagemUtil,
+  gerarMensagemSolicitandoOrderId, // extrairTodosOrderIds foi descontinuada aqui, mas é mantida no utils.js para o parseRawMessage // processarOrderIds foi descontinuada em favor do fluxo no index.js // cortarMensagemUtil foi descontinuada em favor do parseRawMessage
 };
